refactor : search 상수 변수명 변경
