+++ src/constant/search/search.js
@@ -1,9 +1,12 @@
 export const SEARCH = {
-  STATUS: {
-    LOADING: '로딩중',
-    KEYWORD_NONE: '검색어 없음',
+  MESSAGE: {
+    STATUS: {
+      LOADING: '로딩중',
+      KEYWORD_NONE: '검색어 없음',
+    },
   },
   HEADER: '검색 컴포넌트',
   INPUT_HEADER: '검색 하기',
   RECOMMAND_KEYWORD: '추천 검색어',
+  MAX_LIST_LENGTH: 7,
 };
